powerpc: Save/restore PPR for KVM hypercalls
The system call FLIH at 0xc00 unconditionally sets hardware priority to
medium. For hypercalls, this means we lose guest OS priority. The front
end (do_kvm_0x**) to the KVM interrupt handler always assumes that PPR
priority is saved in PACA exception save area, so it copies this to the
kvm_hstate structure. For hypercalls, this would be the saved priority
from any previous exception. Eventually, the guest gets resumed with an
incorrect priority.

The fix is to save the PPR priority in PACA exception save area before
switching HMT priorities in the FLIH so that existing code described above
in the KVM interrupt handler can copy it from there into the VCPU's saved
context.

Signed-off-by: Suresh Warrier <warrier@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>