powerpc/powernv: Enable Offline CPUs to enter deep idle states
The offline cpus should enter deep idle states so as to gain maximum powersavings
when the entire core is offline. To do so the offline path must be made aware
of the available deepest idle state. Hence probe the device tree for the possible
idle states in powernv core code and expose the deepest idle state through flags.
Since the  device tree is probed by the cpuidle driver as well, move the parameters
required to discover the idle states into an appropriate common place to both the
driver and the powernv core code.

Another point is that fastsleep idle state may require workarounds in the kernel to
function properly. This workaround is introduced in the subsequent patches. However
neither the cpuidle driver or the hotplug path need be bothered about this workaround.
They will be taken care of by the core powernv code.

Signed-off-by: Srivatsa S. Bhat <srivatsa.bhat@linux.vnet.ibm.com>
Signed-off-by: Shreyas B. Prabhu <shreyas@linux.vnet.ibm.com>
[ Changelog modified by preeti@linux.vnet.ibm.com ]
Signed-off-by: Preeti U. Murthy <preeti@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>