KVM: PPC: Book3S HV: Don't enable MMU when hwthread_req = 1
Commit 1c7c31659a5f ("powerpc: powernv: Return to cpu offline loop
when finished in guest") made secondary CPU threads jump to
power7_wakeup_loss when they were finished executing in KVM code
rather than going to nap.  However, this is unsafe if the
kvm_hstate.hwthread_req field in the PACA is non-zero.  A non-zero
value in that flag means that the (sub)core is executing in the
guest, or will be shortly, and that the secondary threads should
therefore avoid going to virtual mode in the host kernel (MMU on).

Now, there is the possibility that a secondary CPU thread, one
that doesn't have a VCPU to run, could be woken up by a stray IPI
or interrupt while the core is in a guest.  In that situation we
currently jump to power7_wakeup_loss, which tries to restore the
context where power7_nap() was called.  That is usually a PC value
in pnv_smp_cpu_kill_self with the MMU on.  However, the core is
still in the guest, and the MMU is set to use the guest HPT.

Consequently we take an ISI (instruction storage interrupt) because
the SLB entry for that address specifies a 16MB page (for the linear
mapping), but the guest typically uses 64kB pages for the linear mapping.
With AIL=3, the ISI will be taken at 0xc000_0000_0000_4400 with the
MMU still on.  Consequently we immediately take another ISI, and
continue taking ISIs until the primary thread switches the MMU back
to the host context, at which point we can execute the code at
0xc...4400 and then print an oops like this:

[    1.928634] Unable to handle kernel paging request for instruction fetch
[    1.928638] Faulting instruction address: 0xc000000000004400
[    1.928643] Oops: Kernel access of bad area, sig: 11 [#1]
[    1.928645] SMP NR_CPUS=1024 NUMA PowerNV
[    1.928695] Modules linked in: vhost_net macvtap macvlan xt_CHECKSUM tun bridge 8021q garp stp llc mrp nf_conntrack_netbios_ns nf_conntrack_broadcast ipt_MASQUERADE ip6table_nat nf_nat_ipv6 ip6table_mangle ip6t_REJECT nf_conntrack_ipv6 nf_defrag_ipv6 iptable_nat nf_nat_ipv4 nf_nat iptable_mangle nf_conntrack_ipv4 nf_defrag_ipv4 xt_conntrack nf_conntrack ebtable_filter ebtables ip6table_filter ip6_tables openvswitch ib_ipoib rdma_ucm ib_ucm ib_uverbs ib_umad rdma_cm ib_cm iw_cm ib_addr xfs mlx4_ib ib_sa ib_mad ib_core mlx4_en windfarm_smu_sat i2c_core bnx2x mlx4_core windfarm_pid mdio libcrc32c cxgb4 ses enclosure shpchp raid10 kvm_hv kvm_pr kvm nfsd auth_rpcgss oid_registry nfs_acl lockd sunrpc scsi_dh_alua dm_service_time lpfc ipr scsi_transport_fc scsi_tgt usb_storage dm_multipath
[    1.928700] CPU: 81 PID: 0 Comm: swapper/81 Not tainted 3.10.42-2004.pkvm2_1_1.8.ppc64 #1
[    1.928703] task: c000001fd15249c0 ti: c000001fd15e0000 task.ti: c000001fd15e0000
[    1.928704] NIP: c000000000004400 LR: 00000000000a5408 CTR: 0000000000000000
[    1.928706] REGS: c000001fd15e39f0 TRAP: 0400   Not tainted  (3.10.42-2004.pkvm2_1_1.8.ppc64)
[    1.928712] MSR: 9000000040001030 <SF,HV,ME,IR,DR>  CR: 28004022  XER: 00000000
[    1.928733] CFAR: 000000000002bbc0 SOFTE: 0 \x0aGPR00: 0000000000000000 c000001fd15e3c70 c00000000137de28 0000000000000000 \x0aGPR04: 9000000100001032 c000000000080b54 0000000028004022 0000000000000004 \x0aGPR08: 000000000000000c 4001538fb7000d90 c0000000335f74b4 d000000001e7cb90 \x0aGPR12: 0000000000000500 c000000007ded900 c000001fd15e3f90 0000000000000000 \x0aGPR16: 0000000000000000 0000000000000000 0000000000000000 0000000000000000 \x0aGPR20: c0000000008afa30 0000000000000001 c0000000012a0f17 c000001fd15e0080 \x0aGPR24: 0000000000000008 c000001fd15e0000 c0000000012a1f28 9000000100001032 \x0aGPR28: c000001fd15e0000 0000000000200000 0000000000000051 c000000001418210
[    1.928741] NIP [c000000000004400] instruction_access_relon_pSeries+0x0/0x80
[    1.928745] LR [00000000000a5408] 0xa5408
[    1.928746] Call Trace:
[    1.928751] [c000001fd15e3c70] [c000000000080b54] .pnv_smp_cpu_kill_self+0x84/0x1f0 (unreliable)
[    1.928756] [c000001fd15e3d20] [c000000000046aac] .cpu_die+0x3c/0x60
[    1.928760] [c000001fd15e3d90] [c000000000019328] .arch_cpu_idle_dead+0x18/0x30
[    1.928764] [c000001fd15e3e00] [c0000000001260a0] .cpu_startup_entry+0x2e0/0x300
[    1.928769] [c000001fd15e3ed0] [c000000000883e9c] .start_secondary+0x348/0x350
[    1.928773] [c000001fd15e3f90] [c0000000000097fc] .start_secondary_prolog+0x10/0x14
[    1.928774] Instruction dump:
[    1.928778] e94d0018 614a903c 7d4903a6 4e800420 00000000 00000000 00000000 00000000
[    1.928782] 00000000 00000000 00000000 00000000 <60000000> 7db14ba6 7db04aa6 f92d0080
[    1.928823] ---[ end trace 33ca16e5d89cc980 ]---

To fix this, we spin at low SMT priority while hwthread_req != 0 before
jumping to power7_wakeup_loss.  While spinning, we need to be sensitive
to the possibility of being given a VCPU to run.  Also, we need to recheck
hwthread_req after setting hwthread_state in order to avoid a race, just
like system_reset_pSeries does.  Note that this spinning will only happen
in the rare case when a secondary CPU thread gets woken spuriously.

BZ: 114039

Signed-off-by: Paul Mackerras <paulus@samba.org>