Revert "KVM: PPC: Add hugepage support for IOMMU in-kernel handling"
This reverts commit 8923b54665946aec5d0295405f34d1402a05091d.

The existing real mode acceleration for hugepages keeps the list of pinned
huge pages in KVM - huge pages are pinned in virtual mode and stored
in a list in KVM and realmode handlers can use those pages without
calling get_page() on them.

However this does not work if in-kernel acceleration is not enabled
which is the case for DDW as for DDW we need 64bit version of
KVM_CREATE_SPAPR_TCE and it is not there. As the result, hugepages
get pinned when they are mapped via VFIO ioctls and never get unpinned
as generic IOMMU code expects KVM to do this.

This removes hugepages cache and this disables real mode acceleration
for them. Since huge DMA windows are here, guests with huge pages
will have the entire RAM pinned anyway, no performance decline
is expected.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Signed-off-by: Paul Mackerras <paulus@samba.org>